Revert "Plan 9 from Bell Labs Fourth Edition 2003-03-26"
This reverts commit fb7f0c934c48abaed6040d054ef636408c3c522d.

+++ sys/src/9/mtx/dat.h
@@ -183,7 +183,7 @@ struct
 struct ISAConf {
 	char		*type;
 	ulong	port;
-	int	irq;
+	ulong	irq;
 	ulong	dma;
 	ulong	mem;
 	ulong	size;
